% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwin-data.R
\docType{data}
\name{darwin}
\alias{darwin}
\title{Example data representing the text of Charles Darwin's \emph{On the Origin 
of Species}}
\format{

}
\usage{
data(darwin)
}
\description{
Example data representing the text of Charles Darwin's \emph{On the Origin 
of Species}
}
\examples{
data(darwin)
darwin <- paste0(darwin, collapse = " ")
\donttest{nchar(darwin)}
}
\keyword{datasets}
