% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioiceire-data.R, R/lmm-data.R
\docType{data}
\name{lmm}
\alias{lmm}
\title{Example data representing information about speakers in the Irish component of the International Corpus of English}
\format{
A data frame with 1,332 rows and 9 variables:
\describe{
  \item{Id}{Identifier, factor with 1332 levels}
  \item{FileSpeaker}{Across-file speaker identifier, factor with 1,313 levels}
  \item{File}{File identifier, factor with 300 levels}
  \item{Speaker}{Within-file speaker identifier, factor with 28 levels}
  \item{Zone}{Region where the is situated (IrishRepublic, Mixed, NonCorpusSpeaker, NorthernIreland), factor with 4 levels}
  \item{Date}{Date of recording (1990-1994, 1995-2001, 2002-2005), factor with three levels}
  \item{Gender}{Self reported gender of a speaker (female, male), factor with two levels}
  \item{Age}{Age group a speaker belongs to (0-18, 19-25, 26-33, 34-41, 42-49, 50+), factor with 6 levels}
  \item{Words}{Number of words uttered by a speaker, integer}
}

A data frame with 537 rows and 5 variables:
\describe{
  \item{Date}{Year when a text was produced, integer}
  \item{Genre}{Genre of a text (Bible, Biography, Diary, Education, 
  Fiction, Handbook, History, Law, Philosophy, PrivateLetter,
  PublicLetter, Religion, Science, Sermon, Travel, TrialProceeding), factor with 16 levels}
  \item{Text}{Name of a text as provided by the Penn Parsed Corpora of Historical English, factor with 271 levels}
  \item{Prepositions}{Relative frequency of prepositions (per-1,000-words) in a text, numeric}
  \item{Region}{Region where the text was produced (North, South), factor with two levels}
}
}
\usage{
data(bioiceire)

data(lmm)
}
\description{
Example data representing information about speakers in the 
Irish component of the International Corpus of English including 
the age and gender of speakers, the regional zone where and the date 
when the speakers were recorded, and the words uttered by each speaker. 
In addition, the data set contains information about the files themselves 
such as an identifier, the file name, a combination of file and speaker, 
and a within-file identifier for the speaker.

Example data extracted from the Penn Parsed Corpora of Historical English
which shows the relative frequencies of prepositions in historical texts
with the date when the text was produced, the region where it was produced, 
and the author of the text.
}
\examples{
data(iceirebio)
Gender <- iceirebio$Gender
Words <- iceirebio$Words
\donttest{plot(Gender, Words)}
data(lmm)
Prepositions <- lmm$Prepositions
Date <- lmm$Date
\donttest{plot(Prepositions, Date)}
}
\keyword{datasets}
