% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm-data.R
\docType{data}
\name{lmm}
\alias{lmm}
\title{Example data representing frequencies of prepositions in historical texts}
\format{
A data frame with 537 rows and 5 variables:
\describe{
  \item{Date}{Year when a text was produced, integer}
  \item{Genre}{Genre of a text (Bible, Biography, Diary, Education, 
  Fiction, Handbook, History, Law, Philosophy, PrivateLetter,
  PublicLetter, Religion, Science, Sermon, Travel, TrialProceeding), factor with 16 levels}
  \item{Text}{Name of a text as provided by the Penn Parsed Corpora of Historical English, factor with 271 levels}
  \item{Prepositions}{Relative frequency of prepositions (per-1,000-words) in a text, numeric}
  \item{Region}{Region where the text was produced (North, South), factor with two levels}
}
}
\usage{
data(lmm)
}
\description{
Example data extracted from the Penn Parsed Corpora of Historical English
which shows the relative frequencies of prepositions in historical texts
with the date when the text was produced, the region where it was produced, 
and the author of the text.
}
\examples{
data(lmm)
Prepositions <- lmm$Prepositions
Date <- lmm$Date
\donttest{plot(Prepositions, Date)}
}
\keyword{datasets}
